requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
ibm-cloud-sdk-core==3.19.2
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-platform-services==0.51.2
ibm-vpc==0.21.0
idna==3.6
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rich==13.7.1
six==1.16.0
tzdata==2024.1
urllib3==2.1.0
